Add files via upload
This drivebase is complicated to visualize so I sketched it.  There are a few mistakes and I left out a lot of detail, but I just want to communicate the design. 
Error: the vertical extrusion on the blue/center of the drivebase that the rail is mounted to should be inline with the 2x1 from the drive module.
If we stack the side extrusion from the modules, center, and bumpers, we should: a) add rigidity, and b) allow for movement of bumpers.  I am thinking that pneumatically turned cams from the center and drive module will secure the bumper and add strength to the db.